+++ cartridges/int_klarna_payments/cartridge/scripts/payments/requestBuilder/order.js
@@ -419,7 +419,6 @@
                 .buildOrderLines(order)
                 .buildTotalAmount(order)
                 .buildTotalTax(order)
-                .buildAdditionalCustomerInfo(order)
                 .buildOptions()
                 .buildMerchantInformation(order);
         }